+++ packages/components/src/List/List.tsx
@@ -1,12 +1,11 @@
-import { ReactNode } from 'react';
+import type { ReactNode } from 'react';
 
 import { useClassNames } from '@marigold/system';
-import { HtmlProps } from '@marigold/types';
 
 import { ListContext } from './Context';
 import { ListItem } from './ListItem';
 
-export interface ListProps extends Omit<HtmlProps<'ul'>, 'className'> {
+export interface ListProps {
   variant?: string;
   size?: string;
   /**