+++ themes/theme-docs/src/components/Input.styles.ts
@@ -1,0 +1,28 @@
+// Box
+// ---------------
+export const inputBox = 'border border-border rounded text-text-primary';
+export const inputBackground = 'bg-white';
+
+// Spacing
+// ---------------
+export const xSpacing = {
+  default: 'px-3',
+  small: 'px-2.5',
+};
+
+export const ySpacing = {
+  default: 'py-2',
+  small: 'py-1.5',
+};
+
+export const inputSpacing = {
+  default: `${xSpacing.default} ${ySpacing.default}`,
+  small: `${xSpacing.small} ${ySpacing.small}`,
+};
+
+// States
+// ---------------
+export const inputFocus = '';
+export const inputDisabled = '';
+export const inputError = '';
+export const inputHover = '';